requirements.txt
httpx
pandas
python-dotenv
SQLAlchemy
pytest
ruff
black
